cleanup  checking
